More splitting
Now, both backends can technically be accessed from one variable.
